+++ extensions/healthie/actions/searchAppointments/config/dataPoints.ts
@@ -1,0 +1,12 @@
+import { type DataPointDefinition } from '@awell-health/extensions-core'
+
+export const dataPoints = {
+  results: {
+    key: 'results',
+    valueType: 'number',
+  },
+  appointmentId: {
+    key: 'appointmentId',
+    valueType: 'string',
+  },
+} satisfies Record<string, DataPointDefinition>