+++ extensions/healthie/actions/searchAppointments/index.ts
@@ -1,0 +1,1 @@
+export { searchAppointments } from './searchAppointments'